rename files and its imports to use underscore instead of hyphen
