+++ test/e2e/cypress/e2e/hana_cluster_details.cy.js
@@ -1,4 +1,4 @@
-import * as hanaClusterDetailsPage from '../pageObject/hana-cluster-details-po';
+import * as hanaClusterDetailsPage from '../pageObject/hana_cluster_details_po';
 
 context('HANA cluster details', () => {
   before(() => {
