+++ test/e2e/cypress/pageObject/databases_overview_po.js
@@ -1,5 +1,5 @@
-export * from './base-po.js';
-import * as basePage from './base-po.js';
+export * from './base_po.js';
+import * as basePage from './base_po.js';
 
 // Test Data
 export const hdqDatabase = {
